# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Mobile Shop Management System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /products:
    get:
      operationId: listProducts
      tags: [products]
      summary: List all products
      parameters:
        - name: category
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
        - name: lowStock
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: Products list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
    post:
      operationId: createProduct
      tags: [products]
      summary: Create a product
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
  /products/{id}:
    get:
      operationId: getProduct
      tags: [products]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
    patch:
      operationId: updateProduct
      tags: [products]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
    delete:
      operationId: deleteProduct
      tags: [products]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ─── DEVICES (IMEI / SERIAL) ──────────────────────────────────
...
  /maintenance:
    get:
      operationId: listMaintenance
      tags: [maintenance]
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Maintenance orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MaintenanceOrder"
    post:
      operationId: createMaintenance
      tags: [maintenance]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MaintenanceInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MaintenanceOrder"
  /maintenance/{id}:
    get:
      operationId: getMaintenance
      tags: [maintenance]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Maintenance order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MaintenanceOrder"
    patch:
      operationId: updateMaintenance
      tags: [maintenance]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MaintenanceUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MaintenanceOrder"

  # ─── SOFTWARE SERVICES ────────────────────────────────────────
...
  /reports/sales:
    get:
      operationId: getSalesReport
      tags: [reports]
      parameters:
        - name: from
          in: query
          schema:
            type: string
        - name: to
          in: query
          schema:
            type: string
        - name: groupBy
          in: query
          schema:
            type: string
            enum: [day, month]
      responses:
        "200":
          description: Sales report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SalesReport"
  /reports/profit:
    get:
      operationId: getProfitReport
      tags: [reports]
      parameters:
        - name: from
          in: query
          schema:
            type: string
        - name: to
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Profit report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProfitReport"
...
components:
  schemas:
...
    Product:
      type: object
      required: [id, name, code, category, purchasePrice, salePrice, quantity, alertQuantity]
      properties:
        id:
          type: integer
        name:
          type: string
        nameAr:
          type: ["string", "null"]
        code:
          type: string
        barcode:
          type: ["string", "null"]
        category:
          type: string
        supplierId:
          type: ["integer", "null"]
        supplierName:
          type: ["string", "null"]
        purchasePrice:
          type: number
        salePrice:
          type: number
        minSalePrice:
          type: number
        quantity:
          type: integer
        alertQuantity:
          type: integer
        imageUrl:
          type: ["string", "null"]
        createdAt:
          type: string
    ProductInput:
      type: object
      required: [name, code, category, purchasePrice, salePrice, quantity, alertQuantity]
      properties:
        name:
          type: string
        nameAr:
          type: string
        code:
          type: string
        barcode:
          type: string
        category:
          type: string
        supplierId:
          type: integer
        purchasePrice:
          type: number
        salePrice:
          type: number
        minSalePrice:
          type: number
        quantity:
          type: integer
        alertQuantity:
          type: integer
        imageUrl:
          type: string
    ProductUpdate:
      type: object
      properties:
        name:
          type: string
        nameAr:
          type: string
        barcode:
          type: string
        category:
          type: string
        supplierId:
          type: integer
        purchasePrice:
          type: number
        salePrice:
          type: number
        minSalePrice:
          type: number
        quantity:
          type: integer
        alertQuantity:
          type: integer
        imageUrl:
          type: string

    # ─── DEVICES ─────────────────────────────────────────────────
...
    MaintenanceOrder:
      type: object
      required: [id, ticketNumber, customerName, customerPhone, deviceType, issue, status, createdAt]
      properties:
        id:
          type: integer
        ticketNumber:
          type: string
        customerId:
          type: ["integer", "null"]
        customerName:
          type: string
        customerPhone:
          type: string
        deviceType:
          type: string
        deviceBrand:
          type: ["string", "null"]
        issue:
          type: string
        password:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        status:
          type: string
          enum: [received, inspecting, repairing, waiting_parts, repaired, delivered]
        estimatedCost:
          type: ["number", "null"]
        finalCost:
          type: ["number", "null"]
        technicianId:
          type: ["integer", "null"]
        technicianName:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
    MaintenanceInput:
      type: object
      required: [customerName, customerPhone, deviceType, issue]
      properties:
        customerId:
          type: integer
        customerName:
          type: string
        customerPhone:
          type: string
        deviceType:
          type: string
        deviceBrand:
          type: string
        issue:
          type: string
        password:
          type: string
        notes:
          type: string
        estimatedCost:
          type: number
        technicianId:
          type: integer
    MaintenanceUpdate:
      type: object
      properties:
        status:
          type: string
          enum: [received, inspecting, repairing, waiting_parts, repaired, delivered]
        estimatedCost:
          type: number
        finalCost:
          type: number
        technicianId:
          type: integer
        notes:
          type: string

    # ─── SOFTWARE SERVICES ───────────────────────────────────────
...
    SalesReport:
      type: object
      required: [totalSales, totalRevenue, totalDiscount, chartData]
      properties:
        totalSales:
          type: integer
        totalRevenue:
          type: number
        totalDiscount:
          type: number
        chartData:
          type: array
          items:
            type: object
            required: [period, revenue, count]
            properties:
              period:
                type: string
              revenue:
                type: number
              count:
                type: integer
    ProfitReport:
      type: object
      required: [totalRevenue, totalCost, totalExpenses, netProfit, chartData]
      properties:
        totalRevenue:
          type: number
        totalCost:
          type: number
        totalExpenses:
          type: number
        netProfit:
          type: number
        chartData:
          type: array
          items:
            type: object
            required: [period, profit, revenue, expenses]
            properties:
              period:
                type: string
              profit:
                type: number
              revenue:
                type: number
              expenses:
                type: number
...