# lib/api-zod/src/generated/types/maintenanceInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Mobile Shop Management System API
 * OpenAPI spec version: 0.1.0
 */

export interface MaintenanceInput {
  customerId?: number;
  customerName: string;
  customerPhone: string;
  deviceType: string;
  deviceBrand?: string;
  issue: string;
  password?: string;
  notes?: string;
  estimatedCost?: number;
  technicianId?: number;
}