# lib/api-zod/src/generated/types/maintenanceUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Mobile Shop Management System API
 * OpenAPI spec version: 0.1.0
 */
import type { MaintenanceUpdateStatus } from './maintenanceUpdateStatus';

export interface MaintenanceUpdate {
  status?: MaintenanceUpdateStatus;
  estimatedCost?: number;
  finalCost?: number;
  technicianId?: number;
  notes?: string;
}