# lib/api-zod/src/generated/types/product.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Mobile Shop Management System API
 * OpenAPI spec version: 0.1.0
 */

export interface Product {
  id: number;
  name: string;
  /** @nullable */
  nameAr?: string | null;
  code: string;
  /** @nullable */
  barcode?: string | null;
  category: string;
  /** @nullable */
  supplierId?: number | null;
  /** @nullable */
  supplierName?: string | null;
  purchasePrice: number;
  salePrice: number;
  minSalePrice?: number;
  quantity: number;
  alertQuantity: number;
  /** @nullable */
  imageUrl?: string | null;
  createdAt?: string;
}